summarize-me/requirements.txt
python-dotenv==1.0.1
requests==2.30.0
boto3==1.35.0
weaviate-client==4.8.1
transformers
torch
